refactor: extract AppListBuilder filter/sort/reorder, eliminate repeated casts in generated_form
Phase 2B: Moved filter/sort/reorder logic from apps.dart into a pure AppListBuilder class in lib/components/app_list_builder.dart. Also moved AppsFilter class there to avoid circular imports. Reduces apps.dart build() method by ~150 lines.

Phase 3A: Introduced 'final item = widget.items[r][e]' at the top of the generated_form build loop and replaced ~18 repeated casts with direct property access on the narrowed variable.

Phase 3B: Fixed multi-line cast patterns that survived the replaceAll.
