+++ lib/components/generated_form.dart
@@ -505,16 +505,17 @@ class _GeneratedFormState extends State<GeneratedForm> {
     }
     for (var r = 0; r < formInputs.length; r++) {
       for (var e = 0; e < formInputs[r].length; e++) {
-        String fieldKey = widget.items[r][e].key;
-        if (widget.items[r][e] is GeneratedFormSwitch) {
+        final item = widget.items[r][e];
+        String fieldKey = item.key;
+        if (item is GeneratedFormSwitch) {
           formInputs[r][e] = Row(
             mainAxisAlignment: MainAxisAlignment.spaceBetween,
             children: [
-              Flexible(child: Text(widget.items[r][e].label)),
+              Flexible(child: Text(item.label)),
               const SizedBox(width: 8),
               Switch(
                 value: values[fieldKey],
-                onChanged: (widget.items[r][e] as GeneratedFormSwitch).disabled
+                onChanged: item.disabled
                     ? null
                     : (value) {
                         setState(() {
@@ -525,13 +526,13 @@ class _GeneratedFormState extends State<GeneratedForm> {
               ),
             ],
           );
-        } else if (widget.items[r][e] is GeneratedFormTagInput) {
+        } else if (item is GeneratedFormTagInput) {
           onAddPressed() {
             showDialog<Map<String, dynamic>?>(
               context: context,
               builder: (BuildContext ctx) {
                 return GeneratedFormModal(
-                  title: widget.items[r][e].label,
+                  title: item.label,
                   items: [
                     [GeneratedFormTextField('label', label: tr('label'))],
                   ],
@@ -546,7 +547,7 @@ class _GeneratedFormState extends State<GeneratedForm> {
                   temp ??= {};
                   if (temp[label] == null) {
                     var singleSelect =
-                        (widget.items[r][e] as GeneratedFormTagInput)
+                        item
                             .singleSelect;
                     var someSelected = temp.entries
                         .where((element) => element.value.value)
@@ -569,22 +570,22 @@ class _GeneratedFormState extends State<GeneratedForm> {
               if ((values[fieldKey] as Map<String, MapEntry<int, bool>>?)
                           ?.isNotEmpty ==
                       true &&
-                  (widget.items[r][e] as GeneratedFormTagInput)
+                  item
                       .showLabelWhenNotEmpty)
                 Column(
                   crossAxisAlignment:
-                      (widget.items[r][e] as GeneratedFormTagInput).alignment ==
+                      item.alignment ==
                           WrapAlignment.center
                       ? CrossAxisAlignment.center
                       : CrossAxisAlignment.stretch,
                   children: [
-                    Text(widget.items[r][e].label),
+                    Text(item.label),
                     const SizedBox(height: 8),
                   ],
                 ),
               Wrap(
                 alignment:
-                    (widget.items[r][e] as GeneratedFormTagInput).alignment,
+                    item.alignment,
                 crossAxisAlignment: WrapCrossAlignment.center,
                 children: [
                   ...(values[fieldKey] as Map<String, MapEntry<int, bool>>?)
@@ -615,9 +616,7 @@ class _GeneratedFormState extends State<GeneratedForm> {
                                           .key,
                                       value,
                                     );
-                                    if ((widget.items[r][e]
-                                                as GeneratedFormTagInput)
-                                            .singleSelect &&
+                                    if (item.singleSelect &&
                                         value == true) {
                                       for (var key
                                           in (values[fieldKey]
@@ -707,13 +706,11 @@ class _GeneratedFormState extends State<GeneratedForm> {
                                 });
                               }
 
-                              if ((widget.items[r][e] as GeneratedFormTagInput)
+                              if (item
                                       .deleteConfirmationMessage !=
                                   null) {
                                 var message =
-                                    (widget.items[r][e]
-                                            as GeneratedFormTagInput)
-                                        .deleteConfirmationMessage!;
+                                    item.deleteConfirmationMessage!;
                                 showDialog<Map<String, dynamic>?>(
                                   context: context,
                                   builder: (BuildContext ctx) {
@@ -747,7 +744,7 @@ class _GeneratedFormState extends State<GeneratedForm> {
                             onPressed: onAddPressed,
                             icon: const Icon(Icons.add),
                             label: Text(
-                              (widget.items[r][e] as GeneratedFormTagInput)
+                              item
                                   .label,
                             ),
                           ),
@@ -765,11 +762,11 @@ class _GeneratedFormState extends State<GeneratedForm> {
               ),
             ],
           );
-        } else if (widget.items[r][e] is GeneratedFormSubForm) {
+        } else if (item is GeneratedFormSubForm) {
           List<Widget> subformColumn = [];
           var compact =
-              (widget.items[r][e] as GeneratedFormSubForm).items.length == 1 &&
-              (widget.items[r][e] as GeneratedFormSubForm).items[0].length == 1;
+              item.items.length == 1 &&
+              item.items[0].length == 1;
           for (int i = 0; i < values[fieldKey].length; i++) {
             var internalFormKey = ValueKey(
               generateRandomNumber(
@@ -785,15 +782,14 @@ class _GeneratedFormState extends State<GeneratedForm> {
                   if (!compact) const SizedBox(height: 16),
                   if (!compact)
                     Text(
-                      '${(widget.items[r][e] as GeneratedFormSubForm).label} (${i + 1})',
+                      '${item.label} (${i + 1})',
                       style: const TextStyle(fontWeight: FontWeight.bold),
                     ),
                   GeneratedForm(
                     key: internalFormKey,
                     items:
                         cloneFormItems(
-                              (widget.items[r][e] as GeneratedFormSubForm)
-                                  .items,
+                              item.items,
                             )
                             .map(
                               (x) => x.map((y) {
@@ -833,7 +829,7 @@ class _GeneratedFormState extends State<GeneratedForm> {
                               }
                             : null,
                         label: Text(
-                          '${(widget.items[r][e] as GeneratedFormSubForm).label} (${i + 1})',
+                          '${item.label} (${i + 1})',
                         ),
                         icon: const Icon(Icons.delete_outline_rounded),
                       ),
@@ -853,7 +849,7 @@ class _GeneratedFormState extends State<GeneratedForm> {
                       onPressed: () {
                         values[fieldKey].add(
                           getDefaultValuesFromFormItems(
-                            (widget.items[r][e] as GeneratedFormSubForm).items,
+                            item.items,
                           ),
                         );
                         forceUpdateKeyCount++;
@@ -861,7 +857,7 @@ class _GeneratedFormState extends State<GeneratedForm> {
                       },
                       icon: const Icon(Icons.add),
                       label: Text(
-                        (widget.items[r][e] as GeneratedFormSubForm).label,
+                        item.label,
                       ),
                     ),
                   ),
